BaseDispatcher and BaseHelper were added
