+++ multibotkit/helpers/fb.py
@@ -1,20 +1,11 @@
-from json import JSONDecodeError
-
-import httpx
-from tenacity import (
-    retry,
-    retry_if_exception_type,
-    stop_after_attempt,
-    wait_exponential,
-)
-
+from multibotkit.helpers.base_helper import BaseHelper
 from multibotkit.schemas.fb.outgoing import (
     Message,
     PersistentMenu
 )
 
 
-class FBHelper:
+class FBHelper(BaseHelper):
 
     def __init__(
         self,
@@ -24,30 +15,6 @@ class FBHelper:
     ):
         self.MESSAGES_URL = messages_endpoint + token
         self.PROFILE_URL = profile_endpoint + token
-
-
-    @retry(
-        retry=retry_if_exception_type(httpx.HTTPError)
-        | retry_if_exception_type(JSONDecodeError),
-        reraise=True,
-        stop=stop_after_attempt(5),
-        wait=wait_exponential(multiplier=1, min=4, max=10),
-    )
-    def _perform_sync_request(self, url: str, data: dict):
-        r = httpx.post(url=url, json=data)
-        return r.json()
-
-    @retry(
-        retry=retry_if_exception_type(httpx.HTTPError)
-        | retry_if_exception_type(JSONDecodeError),
-        reraise=True,
-        stop=stop_after_attempt(5),
-        wait=wait_exponential(multiplier=1, min=4, max=10),
-    )
-    async def _perform_async_request(self, url: str, data: dict):
-        async with httpx.AsyncClient() as client:
-            r = await client.post(url=url, json=data)
-        return r.json()
 
 
     def sync_send_message(self, message: Message):
